Store host & port number, and change logic of how user is queried
This now stores the host and port number of the database in the
keyring, so that this can deal with non-default host and port numbers
for the LoCTerms database server.  As part of that, the logic
controlling when the user is asked for the values is now cleaner.
Finally, the -X flag can be used to reset and re-store new values.