Update file headers to new format
I'm settling on a new scheme for file headers for all files that are
not Python module files.  The new scheme uses a comment block and not
documentation strings.  For Python modules, I'm continuing to use a
documentation string under the assumption that this is accessible
within Python and Pythonic tools.